flask-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
sentence-transformers==2.2.2
numpy==1.26.2
pandas==2.1.4
